split simple chart into logical pieces
