+++ cfgov/unprocessed/js/routes/on-demand/simple-chart/analytics.js
